pyproject.toml
[project]
name = "coding-agents-sdlc"
version = "0.1.0"
description = "Automated SDLC Pipeline with AI Agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "langchain>=0.3.7",
    "langchain-openai>=0.2.5",
    "langchain-community>=0.3.5",
    "langgraph>=0.2.45",
    "PyGithub>=2.4.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.35",
    "httpx>=0.27.2",
    "websockets>=13.0",
    "sse-starlette>=2.0.0",
    "aiofiles>=24.0.0",
    "python-multipart>=0.0.9",
    "click>=8.1.7",
    "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.4",
    "black>=24.10.0",
    "mypy>=1.12.1",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
coding-agent = "app.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ============================================================================
# Ruff Configuration
# ============================================================================

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument (needed for callbacks)
    "ARG002", # unused method argument (needed for class methods)
    "TCH001", # type-checking imports (handled separately)
    "TCH002", # type-checking imports (handled separately)
    "TCH003", # type-checking imports (handled separately)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*.py" = ["ARG001", "ARG002"]

# ============================================================================
# Black Configuration
# ============================================================================

[tool.black]
target-version = ["py312"]
line-length = 100
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | build
  | dist
)/
'''

# ============================================================================
# MyPy Configuration
# ============================================================================

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "venv/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "github.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
